% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMPWRFisher.R
\name{fitMPWRFisher}
\alias{fitMPWRFisher}
\title{fitMPWRFisher implements an optimized dynamic programming algorithm to fit a
PWR model.}
\usage{
fitMPWRFisher(X, Y, K, p = 3)
}
\arguments{
\item{X}{Numeric vector of length \emph{m} representing the covariates/inputs
\eqn{x_{1},\dots,x_{m}}.}

\item{Y}{Matrix of size \eqn{(m, d)} representing a \eqn{d} dimension
function of \code{X} observed at points \eqn{1,\dots,m}. \code{Y} is the observed
response/output.}

\item{K}{The number of regimes/segments (PWR components).}

\item{p}{Optional. The order of the polynomial regression. By default, \code{p} is
set at 3.}
}
\value{
fitMPWRFisher returns an object of class \link{ModelMPWR}.
}
\description{
fitMPWRFisher is used to fit a Mulitvariate Piecewise Regression (PWR) model
by maximum-likelihood via an optimized dynamic programming algorithm. The
estimation performed by the dynamic programming algorithm provides an optimal
segmentation of the time series.
}
\details{
fitMPWRFisher function implements an optimized dynamic programming
algorithm of the MPWR model. This function starts with the calculation of
the "cost matrix" then it estimates the transition points given \code{K} the
number of regimes thanks to the method \code{computeDynamicProgram} (method of
the class \link{ParamMPWR}).
}
\examples{
data(toydataset)
x <- toydataset$x
Y <- as.matrix(toydataset[,c("y1", "y2", "y3")])

mpwr <- fitMPWRFisher(X = x, Y = Y, K = 5, p = 1)

mpwr$summary()

mpwr$plot()
}
\seealso{
\link{ModelMPWR}, \link{ParamMPWR}, \link{StatMPWR}
}
