% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMPWR.R
\docType{class}
\name{ModelMPWR-class}
\alias{ModelMPWR-class}
\alias{ModelMPWR}
\title{A Reference Class which represents a fitted MPWR model.}
\description{
ModelMPWR represents an estimated MPWR model.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamMPWR} object. It contains the estimated values
of the parameters.}

\item{\code{stat}}{A \link{StatMPWR} object. It contains all the statistics
associated to the MPWR model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "segmentation"), ...)}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"regressors" = } Polynomial regression components
        (field \code{regressors} of class \link{StatMPWR}).
      \item \code{"segmentation" = } Estimated signal
        (field \code{mean_function} of class \link{StatMPWR}).
    }
  }
  \item{\code{\dots}}{Other graphics parameters.}
}
By default, all the graphs mentioned above are produced.}

\item{\code{summary(digits = getOption("digits"))}}{Summary method.
\describe{
  \item{\code{digits}}{The number of significant digits to use when
    printing.}
}}
}}

\examples{
data(toydataset)
x <- toydataset$x
Y <- as.matrix(toydataset[,c("y1", "y2", "y3")])

mpwr <- fitMPWR(X = x, Y = Y, K = 5, p = 1)

# mpwr is a ModelMPWR object. It contains some methods such as 'summary' and 'plot'
mpwr$summary()
mpwr$plot()

# mpwr has also two fields, stat and param which are reference classes as well

# Value of the objective function:
mpwr$stat$objective

# Parameters of the polynomial regressions:
mpwr$param$beta
}
\seealso{
\link{ParamMPWR}, \link{StatMPWR}
}
