% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPWR-package.R
\docType{package}
\name{MPWR-package}
\alias{MPWR}
\alias{MPWR-package}
\title{MPWR: Multiple Polynomial Piecewise Regression ('MPWR') for Optimal Multivariate Time-Series Segmentation}
\description{
Provides a multiple polynomial piecewise regression model for the 
    optimal segmentation of multivariate time series with regime changes. It 
    uses MLE for the estimation of the regression parameters with dynamic 
    programming for the segmentation. Originally developped and written in 
    'Matlab' by Faicel Chamroukhi <https://github.com/fchamroukhi/MPWR>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fchamroukhi/MPWR_r}
}

}
\author{
\strong{Maintainer}: Florian Lecocq \email{florian.lecocq@outlook.com} (R port) [translator]

Authors:
\itemize{
  \item Faicel Chamroukhi \email{faicel.chamroukhi@unicaen.fr} (0000-0002-5894-3103)
}

}
